[REFACTOR] FoodCard 컴포넌트 삭제 후 home-page로 재배치
