+++ src/pages/home/home-page.tsx
@@ -2,7 +2,6 @@ import { useEffect, useState } from "react";
 import SearchBar from "../../components/common/SearchBar";
 import UpButton from "../../components/common/UpButton";
 import HomeHeader from "../../components/header/HomeHeader";
-import FoodCard from "../../components/home/category/FoodCard";
 import AllergyCuration from "../../components/home/curation/AllergyCuration";
 import MealCuration from "../../components/home/curation/MealCuration";
 import FamilyProfile from "../../components/home/profile/FamilyProfile";
@@ -17,6 +16,11 @@ import { useGetProfile } from "../../hooks/queries/use-get-profile";
 import { useFamilyStore } from "../../stores/use-family-store";
 import { useProfileStore } from "../../stores/use-profile-store";
 import { useMyProfile } from "../../hooks/queries/use-get-my-profile";
+
+import Rice from "../../assets/category/rice.svg";
+import Soup from "../../assets/category/soup.svg";
+import Banchan from "../../assets/category/banchan.svg";
+import Dessert from "../../assets/category/dessert.svg";
 
 const HomePage = () => {
   const navigate = useNavigate();
@@ -139,3 +143,31 @@ const HomePage = () => {
 };
 
 export default HomePage;
+
+const categoryImages: Record<string, string> = {
+  밥: Rice,
+  국: Soup,
+  반찬: Banchan,
+  후식: Dessert,
+};
+
+interface FoodCardProps {
+  name: string;
+  onClick: () => void;
+  isSelected: boolean;
+}
+
+const FoodCard = ({ name, onClick, isSelected }: FoodCardProps) => {
+  const imgSrc = categoryImages[name];
+  return (
+    <div
+      className={`px-2.5 py-2.5 flex flex-col justify-center items-center gap-1 cursor-pointer shrink-0 ${isSelected && "border-b-2 border-primary-700"}`}
+      onClick={onClick}
+    >
+      <img className="w-11 h-11 " src={imgSrc} alt={`${name}`} />
+      <p className="text-center text-neutral-700 text-base font-semibold leading-6">
+        {name}
+      </p>
+    </div>
+  );
+};
